dashboard wip mcp
